Parsex v0.1.0 Parsex

Parser combinator functions.

Functions

andd(parser1, parser2)

Specs

andd(Parsex.Parser.t, Parsex.Parser.t) :: Parsex.Parser.t

Create a parser that succeeds if both of its subparsers succeed

iex> p1 = str("foo")
iex> p2 = str("bar")
iex> foo_and_bar = andd(p1, p2)
iex> foo_and_bar.("foobar")
%Parsex.Parser.Success{result: "foobar", remaining: ""}

iex> p1 = str("foo")
iex> p2 = str("bar")
iex> foo_and_bar = andd(p1, p2)
iex> foo_and_bar.("foobaz")
%Parsex.Parser.Failure{parse_string: "bar", remaining: "baz"}

iex> p1 = str("foo")
iex> p2 = str("bar")
iex> foo_and_bar = andd(p1, p2)
iex> foo_and_bar.("qoobar")
%Parsex.Parser.Failure{parse_string: "foo", remaining: "qoobar"}

orr(parser1, parser2)

Specs

orr(Parsex.Parser.t, Parsex.Parser.t) :: Parsex.Parser.t

Creates a parser that succeeds if one of its subparsers succeeds

iex> p1 = str("foo")
iex> p2 = str("bar")
iex> foo_and_bar = orr(p1, p2)
iex> foo_and_bar.("foo")
%Parsex.Parser.Success{result: "foo", remaining: ""}

iex> p1 = str("foo")
iex> p2 = str("bar")
iex> foo_and_bar = orr(p1, p2)
iex> foo_and_bar.("bar")
%Parsex.Parser.Success{result: "bar", remaining: ""}

iex> p1 = str("foo")
iex> p2 = str("bar")
iex> foo_and_bar = orr(p1, p2)
iex> foo_and_bar.("bazquux")
%Parsex.Parser.Failure{parse_string: "bar", remaining: "bazquux"}

re(regex)

Specs

re(String.t) :: Parsex.Parser.t

Create a parser from a regular expression

iex> foo = re(~r/foo/)
iex> foo.("foo bar")
%Parsex.Parser.Success{result: "foo", remaining: " bar"}

iex> foo = re(~r/foo/)
iex> foo.("bar baz")
%Parsex.Parser.Failure{parse_string: "foo", remaining: "bar baz"}

str(prefix)

Specs

str(String.t) :: Parsex.Parser.t

Create a parser from a String.t.

iex> foo = str("foo")
iex> foo.("foo bar")
%Parsex.Parser.Success{result: "foo", remaining: " bar"}

iex> foo = str("foo")
iex> foo.("baz quux")
%Parsex.Parser.Failure{parse_string: "foo", remaining: "baz quux"}

then(parser, function)

Specs

then(Parsex.Parser.t, (String.t -> term)) :: Parsex.Parser.t

Create a parser that transforms a successful parse

iex> foo = str("foo") |> then(fn(result) -> result <> "bar" end)
iex> foo.("foo")
%Parsex.Parser.Success{result: "foobar", remaining: ""}

iex> foo = str("foo") |> then(fn(result) -> result <> "bar" end)
iex> foo.("quux")
%Parsex.Parser.Failure{parse_string: "foo", remaining: "quux"}

Macros

parser1 <|> parser2

Macro sugar for or/2

iex> p1 = str("foo")
iex> p2 = str("bar")
iex> foo_and_bar = p1 <|> p2
iex> foo_and_bar.("foo")
%Parsex.Parser.Success{result: "foo", remaining: ""}

parser1 <~> parser2

Macro sugar for andd/2

iex> p1 = str("foo")
iex> p2 = str("bar")
iex> foo_and_bar = p1 <~> p2
iex> foo_and_bar.("foobar")
%Parsex.Parser.Success{result: "foobar", remaining: ""}

parser1 ~>> function

Macro sugar for then/2

iex> foo = str("foo") ~>> fn(result) -> result <> "bar" end
iex> foo.("foo")
%Parsex.Parser.Success{result: "foobar", remaining: ""}