Parsex v0.0.1 Parsex

Parser combinator functions.

Functions

and_keep_first(parsers)

Specs

and_keep_first([Parsex.Parser.t]) :: Parsex.Parser.t

AND KEEP FIRST

Creates a logical AND parser from a collection of other parsers. In the case of a success, this parser will return only the result of the first given parser.

iex> and_keep_first([
...> lit("one"),
...> lit("day"),
...> lit("son"),
...> lit("this"),
...> lit("will"),
...> lit("all"),
...> lit("be"),
...> lit("yours")]).("one day son this will all be yours")
{:ok, "", "one"}

and_keep_last(parsers)

Specs

and_keep_last([Parsex.Parser.t]) :: Parsex.Parser.t

AND KEEP LAST

Creates a logical AND parser from a collection of other parsers. In the case of a success, this parser will return only the result of the last given parser.

iex> and_keep_last([
...> lit("one"),
...> lit("day"),
...> lit("son"),
...> lit("this"),
...> lit("will"),
...> lit("all"),
...> lit("be"),
...> lit("ahhh")]).("one day son this will all be ahhh")
{:ok, "", "ahhh"}

iex> and_keep_last([
...> lit("should"),
...> lit("leave"),
...> lit("trailing whitespace ")]).("should leave trailing whitespace ")
{:ok, "", "trailing whitespace "}

and_then(parser, fun)

Specs

and_then(Parsex.Parser.t, (... -> String.t)) :: Parsex.Parser.t

AND THEN

Creates a parser from a given parser and a given function. Applies the function to the result of the given parser and returns the result.

iex> pand([
...> lit("foo"),
...>   pand([lit("bar"), lit("baz")]),
...>   and_then(
...>     lit("quux"),
...>     fn value -> value <> value <> value end)
...>   ]).("foo bar baz quux")
{:ok, "", "foo bar baz quuxquuxquux"}

lit(literal)

Specs

lit(String.t) :: Parsex.Parser.t

STRING LITERAL

Creates a parser from a String literal that succeeds if the given string forms the prefix of the input.

Input is stripped of leading spaces before match.

iex> lit("foo").("foo")
{:ok, "", "foo"}

iex> lit("foo").("bar")
{:error, "literal 'foo' did not match"}

iex> lit("we shall meet").("we shall meet again")
{:ok, " again", "we shall meet"}

pand(parsers)

Specs

pand([Parsex.Parser.t]) :: Parsex.Parser.t

AND

Creates a logical AND parser from a collection of other parsers. This parser will succeed if and only if all of its given parsers succeeds.

iex> pand([lit("foo"), lit("bar")]).("foo bar")
{:ok, "", "foo bar"}

iex> pand([lit("foo"), pand([lit("bar"), lit("baz"), lit("quux")])]).("foo bar baz quux")
{:ok, "", "foo bar baz quux"}

iex> pand([lit("foo"), pand([lit("bar"), lit("baz")])]).("foo bar baz quux")
{:ok, " quux", "foo bar baz"}

iex> pand([lit("foo"), pand([lit("bar"), pregex(~r/w+/)])]).("foo bar")
{:error, "Regex does not match on ''"}

por(parsers)

Specs

por([Parsex.Parser.t]) :: Parsex.Parser.t

OR

Creates a logical OR parser from a collection of other parsers. This parser will succeed if at least one of its given parsers succeeds.

iex> por([lit("foo"), lit("bar"), lit("baz")]).("baz")
{:ok, "", "baz"}

iex> por([lit("foo"), lit("bar"), lit("baz")]).("foo")
{:ok, "", "foo"}

iex> por([lit("foo"), lit("bar"), lit("baz")]).("quux")
{:error, "literal 'baz' did not match"}

pregex(regex)

Specs

pregex(Regex.t) :: Parsex.Parser.t

REGEX

Creates a parser from a Regex literal that succeeds if the given regular expression matches.

Input is stripped of leading spaces before match

iex> pregex(~r/^\w{3} as easy as \d{3}/).("abc as easy as 123")
{:ok, "", "abc as easy as 123"}

replace(parser, replacement)

Specs

replace(Parsex.Parser.t, term) :: Parsex.Parser.t

REPLACE

Creates a parser from a given parser and a given replacement value. Returns the replacement value in place of the result of the given parser.

iex> replace(
...>   lit("han shot first"),
...>   "han shot not even close to first").("han shot first")
{:ok, "", "han shot not even close to first"}

iex> pand([
...> replace(lit("the year of our lord"), "the year of our dark lord"),
...> replace(lit("1776"), "2015")]).("the year of our lord 1776")
{:ok, "", "the year of our dark lord 2015"}

iex> replace(lit("foo"), :bar).("foo")
{:ok, "", "bar"}

Macros

left <|> right

iex> foo_or_bar_or_baz = lit("foo") <|> lit("bar") <|> lit("baz")
iex> foo_or_bar_or_baz.("baz")
{:ok, "", "baz"}

left <~ right

iex> one = lit("one") <~ lit("day son this will all be yours")
...> one.("one day son this will all be yours")
{:ok, "", "one"}

left <~> right

iex> (lit("foo") <~> lit("bar")).("foo bar")
{:ok, "", "foo bar"}

left ~> right

iex> ahhh = lit("one day son this will all be") ~> lit("ahhh")
...> ahhh.("one day son this will all be ahhh")
{:ok, "", "ahhh"}

left ~>> fun

iex> three_times = lit("quux") ~>> fn(value) -> value <> value <> value end
...> three_times.("quux")
{:ok, "", "quuxquuxquux"}